gptfdisk: Autospec creation for version 1.0.1
Add patch to bypass the cgdisk bypass resolution check

This patch allows the cgdisk utility to continue when a small display is
detected after flashing a warning about display issues.

This fixes an issue in the Clear Linux installer on older machines with
lower-resolution diplays (Lenovo ThinkPad S230u) where cgdisk reported
the screen was too small to continue and the user was unable to manually
partition their drive without switching TTYs.

On the mentioned machine this patch fixes the issue, and there are no
breaking issues in the cgdisk utility.